+++ packages/gmana_flutter/lib/spinner/vertical_point.dart
@@ -1,1 +1,0 @@
-export 'package:gmana/math/wave_vertical_offset.dart';